painfull Redux test part 2
